OFConstantString: _cStringLength -> unsigned int.
This field is actually only an unsigned int and not a size_t.

It never caused trouble on 64-bit little endian machines as padding was
added and so the most significant 32 bits were just 0. However, on 64-bit
big endian machines like SPARC64, the most significant 32 bit come
first, and thus the 32 bit size would treated as the 32 bit most
significant bits of a 64 bit value.

This also adds OpenBSD 5.3/SPARC64 to PLATFORMS.md.

+++ src/OFConstantString.h
@@ -33,6 +33,6 @@ extern void *_OFConstantStringClassReference;
 @interface OFConstantString: OFString
 {
 	char *_cString;
-	size_t _cStringLength;
+	unsigned int _cStringLength;
 }
 @end
